+++ src/components/issue/issueItem.js
@@ -2,7 +2,7 @@
 import { BaseComponent } from '../component';
 
 // Templates
-import { getIssueItemTpl } from '../../tpl';
+import { getIssueItemTpl } from '../../template/issue';
 
 export class IssueItem extends BaseComponent {
   constructor(item) {
