+++ src/template/error.js
@@ -1,0 +1,8 @@
+export function getErrorTpl() {
+  /* html */
+  return `
+  <div id="issue-wrapper" class="w-9/12 m-auto min-w-min">
+    페이지를 찾을 수 없습니다.
+  </div>
+  `;
+}
